don't call edm on removeal of gm.. store file in dm object and try delete
git-svn-id: http://svn.vuze.com/public/client@16700 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus3/src/com/aelitis/azureus/util/AdManager.java
@@ -97,10 +97,13 @@ public class AdManager
 		gm.addListener(new GlobalManagerAdapter() {
 			public void downloadManagerRemoved(DownloadManager dm) {
 				adsDMList.remove(dm);
-				try {
-					File asxFile = buildASXFileLocation(dm);
-					asxFile.delete();
-				} catch (Exception e) {
+				File asxFile = (File) dm.getData("ASX");
+				
+				if (asxFile != null) {
+  				try {
+  					asxFile.delete();
+  				} catch (Exception e) {
+  				}
 				}
 			}
 
@@ -111,6 +114,8 @@ public class AdManager
 					hookDM(new DownloadManager[] {
 						dm
 					});
+				} else if (PlatformTorrentUtils.isContentAdEnabled(torrent)) {
+					dm.setData("ASX", buildASXFileLocation(dm));
 				}
 			}
 		}, false);
@@ -434,6 +439,7 @@ public class AdManager
 										playlist.getBytes());
 
 								dm.setData("LastASX", new Long(SystemTime.getCurrentTime()));
+								dm.setData("ASX", asxFile);
 
 								if (l != null) {
 									l.asxCreated(asxFile);
